add test that goes through crawl and then validation
