+++ src/main.rs
@@ -1,18 +1,15 @@
-use anyhow::anyhow;
 use clap::{Parser, Subcommand};
 use tracing_log::LogTracer;
 use tracing_subscriber::{fmt, layer::SubscriberExt, EnvFilter};
 
 use spyglass_lens::LensConfig;
 
-use std::collections::HashSet;
 use std::io;
 use std::path::Path;
 use tokio::runtime;
-use url::Url;
 
-use libnetrunner::archive::Archiver;
-use libnetrunner::{cache_storage_path, Netrunner};
+use libnetrunner::validator::validate_lens;
+use libnetrunner::Netrunner;
 
 const LOG_LEVEL: tracing::Level = tracing::Level::INFO;
 
@@ -47,9 +44,10 @@ enum Commands {
 fn main() -> Result<(), anyhow::Error> {
     // Setup some nice console logging
     let subscriber = tracing_subscriber::registry()
-        .with(EnvFilter::from_default_env()
-            .add_directive(LOG_LEVEL.into())
-            .add_directive(LIB_LOG_LEVEL.parse().expect("invalid log filter"))
+        .with(
+            EnvFilter::from_default_env()
+                .add_directive(LOG_LEVEL.into())
+                .add_directive(LIB_LOG_LEVEL.parse().expect("invalid log filter")),
         )
         .with(fmt::Layer::new().with_writer(io::stdout));
     tracing::subscriber::set_global_default(subscriber).expect("Unable to set a global subscriber");
@@ -102,84 +100,28 @@ fn main() -> Result<(), anyhow::Error> {
             let mut netrunner = Netrunner::new(lens);
             runtime.block_on(netrunner.crawl(false, true))
         }
-        Commands::Validate => {
-            // Check that we can read the WARC file
-            log::info!("Checking for archive file.");
-            let path = cache_storage_path(&lens);
-            match Archiver::read(&path) {
-                Ok(records) => {
-                    let mut failed_validation = false;
-
-                    let urls_txt =
-                        std::fs::read_to_string(path.join("urls.txt")).expect("urls.txt not found");
-                    let expected_urls: HashSet<Url> =
-                        urls_txt.lines()
-                            .flat_map(|x| {
-                                match Url::parse(x) {
-                                    Ok(url) => Some(url),
-                                    Err(err) => {
-                                        failed_validation = true;
-                                        log::error!("Invalid URL in `urls.txt`: <{}> - {}", x, err);
-                                        None
-                                    }
-                                }
-                            })
-                            .collect();
-
-                    let mut zero_len_headers = 0;
-                    let mut zero_len_content = 0;
-
-                    let mut found_urls: HashSet<Url> = HashSet::new();
-                    for rec in &records {
-                        match Url::parse(&rec.url) {
-                            Ok(url) => {
-                                found_urls.insert(url);
-                            },
-                            Err(err) => {
-                                log::error!("Invalid URL in archive: <{}> - {}", rec.url, err);
-                                failed_validation = true;
-                                continue;
-                            }
-                        }
-
-                        if rec.headers.is_empty() {
-                            zero_len_headers += 1;
-                        }
-
-                        if rec.content.is_empty() {
-                            zero_len_content += 1;
-                        }
-                    }
-
-                    if zero_len_headers > 0 {
-                        log::error!("Found {} 0-length headers", zero_len_headers);
-                        failed_validation = true;
-                    }
-
-                    if zero_len_content > 0 {
-                        log::error!("Found {} 0-length content", zero_len_content);
-                        failed_validation = true;
-                    }
-
-                    let missing_urls: Vec<Url> =
-                        expected_urls.difference(&found_urls).cloned().collect();
-
-                    if !missing_urls.is_empty() {
-                        log::error!("{} missing urls", missing_urls.len());
-                        log::error!("{:?}", missing_urls);
-                        failed_validation = true;
-                    }
-
-                    log::info!("Found & validated {} records", records.len());
-                    if failed_validation {
-                        Err(anyhow!("Failed validation"))
-                    } else {
-                        Ok(())
-                    }
-                }
-                Err(e) => Err(e),
-            }
-        }
+        Commands::Validate => validate_lens(&lens),
         _ => Ok(()),
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use libnetrunner::{validator::validate_lens, Netrunner};
+    use spyglass_lens::LensConfig;
+    use std::path::Path;
+
+    #[tokio::test]
+    async fn test_crawl() {
+        let lens_file = "fixtures/test.ron";
+        let lens = LensConfig::from_path(Path::new(&lens_file).to_path_buf())
+            .expect("Unable to load lens file");
+
+        // Test crawling logic
+        let mut netrunner = Netrunner::new(lens.clone());
+        netrunner.crawl(false, true).await.expect("Unable to crawl");
+
+        // Test validation logic
+        validate_lens(&lens).expect("Unable to validate lens");
     }
 }
